move secret-list-header to addon folder
